connectors/rapid7/forgerock/requirements.txt
# This file was autogenerated by uv via the following command:
#    surcom connector codegen
#
# To add a dependency, add it to the `python_dependencies` field in your connector.spec.yaml file and re-run codegen.
#
# To update to use the latest dependencies just re-run the `surcom connector codegen` command.
#
# DO NOT EDIT THIS FILE DIRECTLY.
#


certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
cryptography==46.0.7
    # via -r requirements.in
furl==2.1.4
    # via r7_surcom_api
idna==3.11
    # via requests
orderedmultidict==1.0.2
    # via furl
pycparser==3.0
    # via cffi
pyjwt==2.12.1
    # via -r requirements.in
pyyaml==6.0.3
    # via r7_surcom_api
r7_surcom_api==0.9.11
    # via -r requirements.in
requests==2.33.1
    # via r7_surcom_api
six==1.17.0
    # via
    #   furl
    #   orderedmultidict
tenacity==9.1.4
    # via r7_surcom_api
urllib3==2.6.3
    # via
    #   r7_surcom_api
    #   requests
